Improve BGP functionality
* Test and implement random access to many interesting update message
attributes
* Implement AS path reconstruction
* Implement various AS path iteration patterns (AS_PATH, AS4_PATH and
reconstructed path)
* Implement NEXT_HOP iteration
* Improve performance by eliminating useless memset() over the entire
BGP/MRT packet buffer on close()
* Multiple bugfixes

+++ include/isolario/filterintrin.h
@@ -195,7 +195,7 @@ inline void vm_exec_clrtrie6(filter_vm_t *vm)
 
 inline void vm_exec_settrie(filter_vm_t *vm, int trie)
 {
-    if (unlikely(trie >= vm->ntries))
+    if (unlikely((unsigned int) trie >= vm->ntries))
         vm_abort(vm, VM_TRIE_UNDEFINED);
 
     vm->curtrie = &vm->tries[trie];
@@ -205,7 +205,7 @@ inline void vm_exec_settrie(filter_vm_t *vm, int trie)
 
 inline void vm_exec_settrie6(filter_vm_t *vm, int trie6)
 {
-    if (unlikely(trie6 >= vm->ntries))
+    if (unlikely((unsigned int) trie6 >= vm->ntries))
         vm_abort(vm, VM_TRIE_UNDEFINED);
 
     vm->curtrie6 = &vm->tries[trie6];
